[EVOL]: code review & enhances
