+++ src/v6y-bff/src/resolvers/manager/AuthenticationManager.ts
@@ -1,6 +1,6 @@
 import { ApolloServer } from '@apollo/server';
 import { expressMiddleware } from '@apollo/server/express4';
-import { passportAuthenticate } from '@v6y/commons';
+import { validateCredentials } from '@v6y/commons';
 import BodyParser from 'body-parser';
 import Cors from 'cors';
 import { RequestHandler } from 'express';
@@ -19,7 +19,7 @@ export function buildUserMiddleware(
                     req.body?.operationName !== 'LoginAccount' &&
                     req.body?.operationName !== 'IntrospectionQuery'
                 ) {
-                    const user = await passportAuthenticate(req);
+                    const user = await validateCredentials(req);
                     if (!user) {
                         throw new Error('Unauthorized');
                     }
