+++ src/v6y-commons/src/core/__tests__/AuthenticationHelper-test.ts
@@ -1,8 +1,8 @@
-// PassportUtils.tests.ts
+// AuthenticationHelper.tests.ts
 import { Mock, describe, expect, it, vi } from 'vitest';
 
 import AccountProvider from '../../database/AccountProvider.ts';
-import * as PassportUtils from '../PassportUtils.ts';
+import * as AuthenticationHelper from '../AuthenticationHelper.ts';
 
 vi.mock('../../database/AccountProvider.ts', async () => {
     const actualModule = (await vi.importActual(
@@ -17,9 +17,9 @@ vi.mock('../../database/AccountProvider.ts', async () => {
     };
 });
 
-describe('PassportUtils', () => {
+describe('AuthenticationHelper', () => {
     it('should create jwt options', () => {
-        const result = PassportUtils.createJwtOptions();
+        const result = AuthenticationHelper.createJwtOptions();
         expect(result).toEqual({
             jwtFromRequest: expect.any(Function),
             secretOrKey: expect.any(String),
@@ -28,14 +28,14 @@ describe('PassportUtils', () => {
 
     it('should generate token', () => {
         const account = { _id: 12 };
-        const result = PassportUtils.passportGenerateToken(account);
+        const result = AuthenticationHelper.generateAuthenticationToken(account);
 
         expect(result).toEqual(expect.any(String));
     });
 
     it('should return null if token is invalid', async () => {
         const account = {};
-        const token = PassportUtils.passportGenerateToken(account);
+        const token = AuthenticationHelper.generateAuthenticationToken(account);
 
         const request = {
             cache: 'default' as RequestCache,
@@ -64,7 +64,7 @@ describe('PassportUtils', () => {
             text: async () => '',
         } as unknown as Request;
 
-        const result = await PassportUtils.passportAuthenticate(request);
+        const result = await AuthenticationHelper.validateCredentials(request);
         expect(result).toBe(null);
     });
 
@@ -82,7 +82,7 @@ describe('PassportUtils', () => {
             role: 'ADMIN',
             applications: [1, 2, 3],
         };
-        const token = PassportUtils.passportGenerateToken(account);
+        const token = AuthenticationHelper.generateAuthenticationToken(account);
         const request = {
             cache: 'default' as RequestCache,
             credentials: 'same-origin',
@@ -110,7 +110,7 @@ describe('PassportUtils', () => {
             text: async () => '',
         } as unknown as Request;
 
-        const result = await PassportUtils.passportAuthenticate(request);
+        const result = await AuthenticationHelper.validateCredentials(request);
 
         expect(result).toEqual(account);
     });
